refactor(node): moved changedFilesContainsRegex to utils
